+++ src/main/java/gr/uom/RideOrCry/Repositories/CitizenRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CitizenRepository extends JpaRepository<Citizen, Integer> {
-
+public interface CitizenRepository extends JpaRepository<Citizen, Long> {
     Citizen findByEmailAndPassword(String email, String password);
 }
